refactor(tests): remove swagger-codegen language
