+++ tests/conftest.py
@@ -5,7 +5,7 @@ import pytest
 # indexd_server and indexd_client is needed as fixtures
 from cdisutilstest.code.conftest import indexd_server, indexd_client  # noqa
 from cdisutilstest.code.indexd_fixture import clear_database
-import swagger_client
+
 
 from indexd import auth
 import importlib
@@ -34,51 +34,6 @@ def user(app):
         "Content-Type": "application/json",
     }
     app.auth.delete("test")
-
-
-@pytest.fixture
-def swg_config(indexd_client):  # noqa
-    config = swagger_client.Configuration()
-    config.host = indexd_client.url
-    config.username = indexd_client.auth[0]
-    config.password = indexd_client.auth[1]
-    yield config
-
-
-@pytest.fixture
-def swg_index_client(swg_config):
-    api = swagger_client.IndexApi(swagger_client.ApiClient(swg_config))
-    yield api
-
-
-@pytest.fixture
-def swg_global_client(swg_config):
-    api = swagger_client.GlobalApi(swagger_client.ApiClient(swg_config))
-    yield api
-
-
-@pytest.fixture
-def swg_alias_client(swg_config):
-    api = swagger_client.AliasApi(swagger_client.ApiClient(swg_config))
-    yield api
-
-
-@pytest.fixture
-def swg_dos_client(swg_config):
-    api = swagger_client.DOSApi(swagger_client.ApiClient(swg_config))
-    yield api
-
-
-@pytest.fixture
-def swg_query_client(swg_config):
-    api = swagger_client.QueryApi(swagger_client.ApiClient(swg_config))
-    yield api
-
-
-@pytest.fixture
-def swg_bulk_client(swg_config):
-    api = swagger_client.BulkApi(swagger_client.ApiClient(swg_config))
-    yield api
 
 
 @pytest.fixture
