+++ tests/test_client.py
@@ -2,9 +2,7 @@ import json
 
 import pytest
 
-from tests.util import assert_blank
 from indexd.index.blueprint import ACCEPTABLE_HASHES
-from swagger_client.rest import ApiException
 
 
 def get_doc(
@@ -27,19 +25,24 @@ def get_doc(
     return doc
 
 
-def test_index_list(swg_index_client):
-    r = swg_index_client.list_entries()
-    assert r.records == []
+def test_index_list(client):
+    res = client.get("/index/")
+    rec = res.json
+    assert rec["records"] == []
 
 
-def test_index_list_with_params(swg_index_client):
-    data = get_doc()
-    data["urls"] = ["s3://endpointurl/bucket_2/key_2", "s3://anotherurl/bucket_2/key_2"]
-    data["urls_metadata"] = {
+def test_index_list_with_params(client, user):
+    data1 = get_doc()
+    data1["urls"] = [
+        "s3://endpointurl/bucket_2/key_2",
+        "s3://anotherurl/bucket_2/key_2",
+    ]
+    data1["urls_metadata"] = {
         "s3://endpointurl/bucket_2/key_2": {"state": "error", "other": "xxx"},
         "s3://anotherurl/bucket_2/key_2": {"state": "error", "other": "xxx"},
     }
-    r_1 = swg_index_client.add_entry(data)
+    res_1 = client.post("/index/", json=data1, headers=user)
+    rec_1 = res_1.json
 
     data2 = get_doc()
     data2["metadata"] = {"project_id": "other-project", "state": "abc", "other": "xxx"}
@@ -47,248 +50,313 @@ def test_index_list_with_params(swg_index_client):
     data2["urls_metadata"] = {
         "s3://endpointurl/bucket/key_2": {"state": "error", "other": "xxx"}
     }
-    r_2 = swg_index_client.add_entry(data2)
-    r = swg_index_client.list_entries(metadata="project_id:bpa-UChicago")
-    ids = [record.did for record in r.records]
-    assert r_1.did in ids
+    res_2 = client.post("/index/", json=data2, headers=user)
+    rec_2 = res_2.json
 
-    r = swg_index_client.list_entries(metadata="project_id:other-project")
-    ids = [record.did for record in r.records]
-    assert r_2.did in ids
+    data1_by_md = client.get("/index/?metadata=project_id:bpa-UChicago")
+    data1_list = data1_by_md.json
+    ids = [record["did"] for record in data1_list["records"]]
+    assert rec_1["did"] in ids
 
-    r = swg_index_client.list_entries(hash="md5:8b9942cf415384b27cadf1f4d2d682e5")
-    ids = [record.did for record in r.records]
-    assert r_1.did in ids
-    assert r_2.did in ids
+    data2_by_md = client.get("/index/?metadata=project_id:other-project")
+    data2_list = data2_by_md.json
+    ids = [record["did"] for record in data2_list["records"]]
+    assert rec_2["did"] in ids
 
-    r = swg_index_client.list_entries(ids=",".join(ids))
+    data_by_hash = client.get("/index/?hash=md5:8b9942cf415384b27cadf1f4d2d682e5")
+    data_list_all = data_by_hash.json
+    ids = [record["did"] for record in data_list_all["records"]]
+    assert rec_1["did"] in ids
+    assert rec_2["did"] in ids
 
-    ids = [record.did for record in r.records]
-    assert r_1.did in ids
-    assert r_2.did in ids
+    idslist = ",".join(ids)
+    data_by_ids = client.get("/index/?ids=" + idslist)
+    data_list_all = data_by_ids.json
 
-    r = swg_index_client.list_entries(limit=2)
-    assert len(r.records) == 2
+    ids = [record["did"] for record in data_list_all["records"]]
+    assert rec_1["did"] in ids
+    assert rec_2["did"] in ids
+
+    data_with_limit = client.get("/index/?limit=2")
+    data_list_limit = data_with_limit.json
+    assert len(data_list_limit["records"]) == 2
 
     param = {"bucket": {"state": "error", "other": "xxx"}}
 
-    r = swg_index_client.list_entries(limit=2, urls_metadata=json.dumps(param))
-    assert len(r.records) == 2
+    data_by_url_md = client.get("/index/?" + json.dumps(param) + "limit=2")
+    data_list_limit = data_by_url_md.json
+    assert len(data_list_limit["records"]) == 2
 
 
-def test_index_list_with_params_negate(swg_index_client):
+def test_index_list_with_params_negate(client, user):
     data = get_doc()
-    r_1 = swg_index_client.add_entry(data)
+    res_1 = client.post("/index/", json=data, headers=user)
+    rec_1 = res_1.json
 
     data["metadata"] = {"testkey": "test", "project_id": "negate-project"}
-    r_2 = swg_index_client.add_entry(data)
+    res_2 = client.post("/index/", json=data, headers=user)
+    rec_2 = res_2.json
 
     data["urls"] = ["s3://endpointurl/bucket_2/key_2", "s3://anotherurl/bucket_2/key_2"]
     data["urls_metadata"] = {"s3://endpointurl/bucket_2/key_2": {"state": "error"}}
-    r_3 = swg_index_client.add_entry(data)
+    res_3 = client.post("/index/", json=data, headers=user)
+    rec_3 = res_3.json
 
     data["urls"] = ["s3://endpointurl/bucket_2/key_2"]
     data["urls_metadata"] = {
         "s3://endpointurl/bucket_2/key_2": {"no_state": "uploaded"}
     }
-    r_4 = swg_index_client.add_entry(data)
+    res_4 = client.post("/index/", json=data, headers=user)
+    rec_4 = res_4.json
 
     data["urls"] = ["s3://anotherurl/bucket/key"]
     data["urls_metadata"] = {"s3://anotherurl/bucket/key": {"state": "error"}}
-    r_5 = swg_index_client.add_entry(data)
+    res_5 = client.post("/index/", json=data, headers=user)
+    rec_5 = res_5.json
 
     negate_params = {"metadata": {"testkey": ""}}
-    r = swg_index_client.list_entries(negate_params=json.dumps(negate_params))
-    ids = {record.did for record in r.records}
-    assert {r_1.did} == ids
+    data_neg_param = client.get("/index/?negate_params=" + json.dumps(negate_params))
+    data_list = data_neg_param.json
+    ids = {record["did"] for record in data_list["records"]}
+    assert {rec_1["did"]} == ids
 
     negate_params = {"metadata": {"project_id": "bpa-UChicago"}}
-    r = swg_index_client.list_entries(negate_params=json.dumps(negate_params))
-    ids = {record.did for record in r.records}
-    assert {r_2.did, r_3.did, r_4.did, r_5.did} == ids
+    data_neg_param = client.get("/index/?negate_params=" + json.dumps(negate_params))
+    data_list = data_neg_param.json
+    ids = {record["did"] for record in data_list["records"]}
+    assert {rec_2["did"], rec_3["did"], rec_4["did"], rec_5["did"]} == ids
 
     # negate url
     negate_params = {"urls": ["s3://endpointurl/bucket_2/key_2"]}
-    r = swg_index_client.list_entries(negate_params=json.dumps(negate_params))
-    ids = {record.did for record in r.records}
-    assert ids == {r_1.did, r_2.did, r_5.did}
+    data_neg_param = client.get("/index/?negate_params=" + json.dumps(negate_params))
+    data_list = data_neg_param.json
+    ids = {record["did"] for record in data_list["records"]}
+    assert ids == {rec_1["did"], rec_2["did"], rec_5["did"]}
 
     # negate url key
     negate_params = {"urls_metadata": {"s3://endpointurl/": {}}}
-    r = swg_index_client.list_entries(negate_params=json.dumps(negate_params))
-    ids = {record.did for record in r.records}
-    assert ids == {r_5.did}
+    data_neg_param = client.get("/index/?negate_params=" + json.dumps(negate_params))
+    data_list = data_neg_param.json
+    ids = {record["did"] for record in data_list["records"]}
+    assert ids == {rec_5["did"]}
 
     negate_params = {"urls_metadata": {"s3://endpointurl/": {}, "s3://anotherurl/": {}}}
-    r = swg_index_client.list_entries(negate_params=json.dumps(negate_params))
-    ids = {record.did for record in r.records}
+    data_neg_param = client.get("/index/?negate_params=" + json.dumps(negate_params))
+    data_list = data_neg_param.json
+    ids = {record["did"] for record in data_list["records"]}
     assert ids == set()
 
     # negate url_metadata key
     negate_params = {
         "urls_metadata": {"s3://endpointurl/": {"state": ""}, "s3://anotherurl/": {}}
     }
-    r = swg_index_client.list_entries(negate_params=json.dumps(negate_params))
-    ids = {record.did for record in r.records}
-    assert ids == {r_4.did}
+    data_neg_param = client.get("/index/?negate_params=" + json.dumps(negate_params))
+    data_list = data_neg_param.json
+    ids = {record["did"] for record in data_list["records"]}
+    assert ids == {rec_4["did"]}
 
     # negate url_metadata value
     negate_params = {"urls_metadata": {"s3://endpointurl/": {"state": "uploaded"}}}
-    r = swg_index_client.list_entries(negate_params=json.dumps(negate_params))
-    ids = {record.did for record in r.records}
-    assert ids == {r_3.did, r_4.did, r_5.did}
+    data_neg_param = client.get("/index/?negate_params=" + json.dumps(negate_params))
+    data_list = data_neg_param.json
+    ids = {record["did"] for record in data_list["records"]}
+    assert ids == {rec_3["did"], rec_4["did"], rec_5["did"]}
 
 
-def test_list_entries_with_uploader(swg_index_client):
+def test_list_entries_with_uploader(client, user):
     """
     Test that return a list of record given uploader
     """
     data = get_doc()
     data["uploader"] = "uploader_1"
-    swg_index_client.add_entry(data)
+    res_1 = client.post("/index/", json=data, headers=user)
+    rec_1 = res_1.json
 
     data = get_doc()
     data["uploader"] = "uploader_123"
-    r2 = swg_index_client.add_entry(data)
+    res_2 = client.post("/index/", json=data, headers=user)
+    rec_2 = res_2.json
 
     data = get_doc()
     data["uploader"] = "uploader_123"
-    r3 = swg_index_client.add_entry(data)
+    res_3 = client.post("/index/", json=data, headers=user)
+    rec_3 = res_3.json
 
-    r = swg_index_client.list_entries(uploader="uploader_123")
-    assert len(r.records) == 2
-    assert {r2.did, r3.did} == {r.records[0].did, r.records[1].did}
-    assert r.records[0].uploader == "uploader_123"
-    assert r.records[1].uploader == "uploader_123"
+    data_grab = client.get("/index/?uploader=uploader_123")
+    data_list = data_grab.json
+    assert len(data_list["records"]) == 2
+    assert {rec_2["did"], rec_3["did"]} == {
+        data_list["records"][0]["did"],
+        data_list["records"][1]["did"],
+    }
+    assert data_list["records"][0]["uploader"] == "uploader_123"
+    assert data_list["records"][1]["uploader"] == "uploader_123"
 
 
-def test_list_entries_with_uploader_wrong_uploader(swg_index_client):
+def test_list_entries_with_uploader_wrong_uploader(client, user):
     """
     Test that returns no record due to wrong uploader id
     """
     data = get_doc()
     data["uploader"] = "uploader_1"
-    r = swg_index_client.add_entry(data)
+    res = client.post("/index/", json=data, headers=user)
 
     data = get_doc()
     data["uploader"] = "uploader_123"
-    r = swg_index_client.add_entry(data)
+    res = client.post("/index/", json=data, headers=user)
 
     data = get_doc()
     data["uploader"] = "uploader_123"
-    r = swg_index_client.add_entry(data)
+    res = client.post("/index/", json=data, headers=user)
 
-    r = swg_index_client.list_entries(uploader="wrong_uploader")
-    assert len(r.records) == 0
+    data_grab = client.get("/index/?uploader=wrong_uploader")
+    data_list = data_grab.json
+    assert len(data_list["records"]) == 0
 
 
-def test_create_blank_record(swg_index_client):
+def test_create_blank_record(client, user):
     """
     Test that new blank records only contain the uploader
     and optionally file_name fields: test without file name
     """
 
     doc = {"uploader": "uploader_123"}
-    r = swg_index_client.create_blank_entry(doc)
-    assert r.did
-    assert r.rev
-    assert r.baseid
+    res = client.post("/index/blank/", json=doc, headers=user)
+    rec = res.json
+    assert rec["did"]
+    assert rec["rev"]
+    assert rec["baseid"]
 
-    r = swg_index_client.list_entries(uploader="uploader_123")
-    assert r.records[0].uploader == "uploader_123"
-    assert not r.records[0].file_name
-    assert_blank(r)
+    res = client.get("/index/?uploader=uploader_123")
+    rec = res.json
+    assert rec["records"][0]["uploader"] == "uploader_123"
+    assert not rec["records"][0]["file_name"]
+
+    # statements below replace assert_blank
+    assert rec["records"][0]["baseid"]
+    assert rec["records"][0]["did"]
+    assert not rec["records"][0]["size"]
+    assert not rec["records"][0]["acl"]
+    assert not rec["records"][0]["authz"]
+    assert not rec["records"][0]["hashes"]
 
 
-def test_create_blank_record_with_file_name(swg_index_client):
+def test_create_blank_record_with_file_name(client, user):
     """
     Test that new blank records only contain the uploader
     and optionally file_name fields: test with file name
     """
 
     doc = {"uploader": "uploader_321", "file_name": "myfile.txt"}
-    r = swg_index_client.create_blank_entry(doc)
-    assert r.did
-    assert r.rev
-    assert r.baseid
+    res = client.post("/index/blank/", json=doc, headers=user)
+    rec = res.json
+    assert rec["did"]
+    assert rec["rev"]
+    assert rec["baseid"]
 
-    r = swg_index_client.list_entries(uploader="uploader_321")
-    assert r.records[0].uploader == "uploader_321"
-    assert r.records[0].file_name == "myfile.txt"
-    assert_blank(r)
+    res = client.get("/index/?uploader=uploader_321")
+    rec = res.json
+    assert rec["records"][0]["uploader"] == "uploader_321"
+    assert rec["records"][0]["file_name"] == "myfile.txt"
+
+    # statements below replace asser_blank
+    assert rec["records"][0]["baseid"]
+    assert rec["records"][0]["did"]
+    assert not rec["records"][0]["size"]
+    assert not rec["records"][0]["acl"]
+    assert not rec["records"][0]["authz"]
+    assert not rec["records"][0]["hashes"]
 
 
-def test_fill_size_n_hash_for_blank_record(swg_index_client):
+def test_fill_size_n_hash_for_blank_record(client, user):
     """
     Test that can fill size and hashes for empty record
     """
     doc = {"uploader": "uploader_123"}
 
-    r = swg_index_client.create_blank_entry(doc)
-    assert r.did
-    assert r.rev
+    res = client.post("/index/blank/", json=doc, headers=user)
+    print(res.status_code)
+    rec = res.json
+    print(rec)
+    assert rec["did"]
+    assert rec["rev"]
 
-    did, rev = r.did, r.rev
+    did, rev = rec["did"], rec["rev"]
     updated = {"size": 10, "hashes": {"md5": "8b9942cf415384b27cadf1f4d2d981f5"}}
 
-    r = swg_index_client.update_blank_entry(did, rev=rev, body=updated)
-    assert r.did == did
-    assert r.rev != rev
+    res = client.put("/index/blank/" + did + "?rev=" + rev, headers=user, json=updated)
+    rec = res.json
+    assert rec["did"] == did
+    assert rec["rev"] != rev
 
-    r = swg_index_client.get_entry(did)
-    assert r.size == 10
-    assert r.hashes.md5 == "8b9942cf415384b27cadf1f4d2d981f5"
+    res = client.get("/index/" + did)
+    rec = res.json
+    assert rec["size"] == 10
+    assert rec["hashes"]["md5"] == "8b9942cf415384b27cadf1f4d2d981f5"
 
 
-def test_get_empty_acl_authz_record(swg_index_client):
+def test_get_empty_acl_authz_record(client, user):
     """
     Test that can get a list of empty acl/authz given uploader
     """
     doc = get_doc()
-    r = swg_index_client.add_entry(doc)
+    res_1 = client.post("/index/", json=doc, headers=user)
+    rec_1 = res_1.json
 
     doc = {"uploader": "uploader_123"}
-    r2 = swg_index_client.create_blank_entry(doc)
+    res_2 = client.post("/index/blank/", json=doc, headers=user)
+    rec_2 = res_2.json
 
     doc = {"uploader": "uploader_123"}
-    r3 = swg_index_client.create_blank_entry(doc)
+    res_3 = client.post("/index/blank/", json=doc, headers=user)
+    rec_3 = res_3.json
 
-    r = swg_index_client.list_entries()
-    assert len(r.records) == 3
+    data_grab = client.get("/index/")
+    data_list = data_grab.json
+    assert len(data_list["records"]) == 3
 
-    r = swg_index_client.list_entries(uploader="uploader_123", acl="null", authz="null")
+    data_by_acl_authz = client.get("/index/?uploader=uploader_123&acl=null&authz=null")
+    data_list = data_by_acl_authz.json
 
-    assert len(r.records) == 2
-    assert {r2.did, r3.did} == {r.records[0].did, r.records[1].did}
-    assert r.records[0].acl == []
-    assert r.records[1].acl == []
-    assert r.records[0].authz == []
-    assert r.records[1].authz == []
+    assert len(data_list["records"]) == 2
+    assert {rec_2["did"], rec_3["did"]} == {
+        data_list["records"][0]["did"],
+        data_list["records"][1]["did"],
+    }
+    assert data_list["records"][0]["acl"] == []
+    assert data_list["records"][1]["acl"] == []
+    assert data_list["records"][0]["authz"] == []
+    assert data_list["records"][1]["authz"] == []
 
 
-def test_get_empty_acl_authz_record_after_fill_size_n_hash(swg_index_client):
+def test_get_empty_acl_authz_record_after_fill_size_n_hash(client, user):
     """
     Test create blank record -> fill hash and size -> get record with empty or none
     acl/authz
     """
     # create the first blank record, update size, hashes and acl/authz
     doc = {"uploader": "uploader_123"}
-    r1 = swg_index_client.create_blank_entry(doc)
+    res_1 = client.post("/index/blank/", json=doc, headers=user)
+    rec_1 = res_1.json
     updated = {"size": 10, "hashes": {"md5": "8b9942cf415384b27cadf1f4d2d981f5"}}
-    did1 = r1.did
-    r1 = swg_index_client.update_blank_entry(r1.did, rev=r1.rev, body=updated)
-    r1 = swg_index_client.update_entry(
-        r1.did, rev=r1.rev, body={"acl": ["read"], "authz": ["read"]}
-    )
-    r1 = swg_index_client.get_entry(r1.did)
-    assert r1.acl == ["read"]
-    assert r1.authz == ["read"]
-    assert r1.did == did1
+    did_1 = rec_1["did"]
+    res_2 = client.put("/index/blank/" + did_1, headers=user, json=updated)
+    rec_2 = res_2.json
+    rev = rec_1["rev"]
+    body = {"acl": ["read"], "authz": ["read"]}
+    res_1 = client.put("/index/" + did_1 + "?rev=" + rev, headers=user, json=body)
+    rec_1 = res_1.json
+    res_1 = client.get("/index/" + rec_1["did"])
+    rec_1 = res_1.json
+    assert rec_1["acl"] == ["read"]
+    assert rec_1["authz"] == ["read"]
+    assert rec_1["did"] == did_1
 
     # create the second blank record, only update size hashes and urls
     doc = {"uploader": "uploader_123"}
-    r2 = swg_index_client.create_blank_entry(doc)
-    did2 = r2.did
+    res_2 = client.post("/index/blank/", json=doc, headers=user)
+    rec_2 = res_2.json
+    did_2 = rec_2["did"]
     updated = {
         "size": 4,
         "hashes": {"md5": "1b9942cf415384b27cadf1f4d2d981f5"},
@@ -297,42 +365,57 @@ def test_get_empty_acl_authz_record_after_fill_size_n_hash(swg_index_client):
 
     # create the second blank record, only update size hashes and urls
     doc = {"uploader": "uploader_123"}
-    r3 = swg_index_client.create_blank_entry(doc)
-    did3 = r3.did
+    res_3 = client.post("/index/blank/", json=doc, headers=user)
+    rec_3 = res_3.json
+    did_3 = rec_3["did"]
     updated = {
         "size": 4,
         "hashes": {"md5": "1b9942cf415384b27cadf1f4d2d981f5"},
         "urls": ["s3://example/2"],
     }
-    swg_index_client.update_blank_entry(r3.did, rev=r3.rev, body=updated)
+    client.put(
+        "/index/blank/" + rec_3["did"] + "?rev=" + rec_3["rev"],
+        json=updated,
+        headers=user,
+    )
 
-    r = swg_index_client.list_entries(uploader="uploader_123")
-    assert len(r.records) == 3
+    res = client.get("/index/?uploader=uploader_123")
+    rec = res.json
+    print(rec)
+    assert len(rec["records"]) == 3
 
-    r = swg_index_client.list_entries(uploader="uploader_123", acl="read")
-    assert len(r.records) == 1
-    assert r.records[0].did == r1.did
+    res = client.get("/index/?uploader=uploader_123&acl=read")
+    rec = res.json
+    assert len(rec["records"]) == 1
+    assert rec["records"][0]["did"] == rec_1["did"]
 
-    r = swg_index_client.list_entries(uploader="uploader_123", acl="write")
-    assert len(r.records) == 0
+    res = client.get("/index/?uploader=uploader_123&acl=write")
+    rec = res.json
+    assert len(rec["records"]) == 0
 
-    r = swg_index_client.list_entries(uploader="uploader_123", acl="null")
-    assert len(r.records) == 2
-    assert {r.records[0].did, r.records[1].did} == {did2, did3}
+    res = client.get("/index/?uploader=uploader_123&acl=null")
+    rec = res.json
+    assert len(rec["records"]) == 2
+    assert {rec["records"][0]["did"], rec["records"][1]["did"]} == {did_2, did_3}
 
 
-def test_urls_metadata(swg_index_client):
+def test_urls_metadata(client, user):
     data = get_doc(has_urls_metadata=True)
-    result = swg_index_client.add_entry(data)
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
 
-    doc = swg_index_client.get_entry(result.did)
-    assert doc.urls_metadata == data["urls_metadata"]
+    res_2 = client.get("/index/" + rec["did"])
+    rec_2 = res_2.json
+    assert rec_2["urls_metadata"] == data["urls_metadata"]
 
     updated = {"urls_metadata": {data["urls"][0]: {"test": "b"}}}
-    swg_index_client.update_entry(doc.did, rev=doc.rev, body=updated)
+    client.put(
+        "/index/" + rec_2["did"] + "?rev=" + rec_2["rev"], json=updated, headers=user
+    )
 
-    doc = swg_index_client.get_entry(result.did)
-    assert doc.urls_metadata == updated["urls_metadata"]
+    res_3 = client.get("/index/" + rec["did"])
+    rec_3 = res_3.json
+    assert rec_3["urls_metadata"] == updated["urls_metadata"]
 
 
 @pytest.mark.parametrize(
@@ -381,7 +464,7 @@ def test_urls_metadata(swg_index_client):
     ],
 )
 def test_urls_metadata_partial_match(
-    swg_index_client, doc_urls, urls_meta, params, expected
+    client, doc_urls, urls_meta, params, expected, user
 ):
     url_doc_mapping = {}
     for url_group in doc_urls:
@@ -391,68 +474,82 @@ def test_urls_metadata_partial_match(
         for url in url_group:
             data["urls_metadata"][url] = urls_meta[url]
 
-        record = swg_index_client.add_entry(data)
+        res = client.post("/index/", json=data, headers=user)
+        rec = res.json
         for url in url_group:
-            url_doc_mapping[url] = record
+            url_doc_mapping[url] = rec
 
-    docs = swg_index_client.list_entries(urls_metadata=json.dumps(params))
+    res = client.get("/index/?urls_metadata=" + json.dumps(params))
+    rec = res.json
 
-    ids = {r.did for r in docs.records}
-    assert ids == {url_doc_mapping[url].did for url in expected}
+    ids = {r["did"] for r in rec["records"]}
+    assert ids == {url_doc_mapping[url]["did"] for url in expected}
 
 
-def test_get_urls(swg_index_client, swg_global_client):
+def test_get_urls(client, user):
     data = get_doc(has_urls_metadata=True)
-    result = swg_index_client.add_entry(data)
+    response = client.post("/index/", json=data, headers=user)
+    record = response.json
 
-    result = swg_global_client.list_urls(ids=result.did)
+    response = client.get("/urls/?ids=" + record["did"])
+    record = response.json
     url = data["urls"][0]
-    assert result.urls[0].url == url
-    assert result.urls[0].metadata == data["urls_metadata"][url]
+    assert record["urls"][0]["url"] == url
+    assert record["urls"][0]["metadata"] == data["urls_metadata"][url]
 
 
-def test_index_create(swg_index_client):
+def test_index_create(client, user):
     data = get_doc(has_baseid=True)
 
-    result = swg_index_client.add_entry(data)
-    assert result.did
-    assert result.baseid == data["baseid"]
-    r = swg_index_client.get_entry(result.did)
-    assert r.acl == []
-    assert r.authz == []
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    assert rec["did"]
+    assert rec["baseid"] == data["baseid"]
+    res = client.get("/index/" + rec["did"])
+    rec = res.json
+    assert rec["acl"] == []
+    assert rec["authz"] == []
 
 
-def test_index_get(swg_index_client):
+def test_index_get(client, user):
     data = get_doc(has_baseid=True)
 
-    result = swg_index_client.add_entry(data)
-    r = swg_index_client.get_entry(result.did)
-    r2 = swg_index_client.get_entry(result.baseid)
-    assert r.did == result.did
-    assert r2.did == result.did
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    res_1 = client.get("/index/" + rec["did"])
+    rec_1 = res_1.json
+    res_2 = client.get("/index/" + rec["baseid"])
+    rec_2 = res_2.json
+    assert rec_1["did"] == rec["did"]
+    assert rec_2["did"] == rec["did"]
 
 
-def test_index_prepend_prefix(swg_index_client):
+def test_index_prepend_prefix(client, user):
     data = get_doc()
 
-    result = swg_index_client.add_entry(data)
-    r = swg_index_client.get_entry(result.did)
-    assert r.did == result.did
-    assert r.did.startswith("testprefix:")
+    res_1 = client.post("/index/", json=data, headers=user)
+    rec_1 = res_1.json
+    res_2 = client.get("/index/" + rec_1["did"])
+    rec_2 = res_2.json
+
+    assert rec_1["did"] == rec_2["did"]
+    assert rec_2["did"].startswith("testprefix:")
 
 
-def test_index_get_with_baseid(swg_index_client):
+def test_index_get_with_baseid(client, user):
     data1 = get_doc(has_baseid=True)
-    swg_index_client.add_entry(data1)
+    client.post("/index/", json=data1, headers=user)
 
     data2 = get_doc(has_baseid=True)
-    r2 = swg_index_client.add_entry(data2)
+    res_1 = client.post("/index/", json=data2, headers=user)
+    rec_1 = res_1.json
 
-    r = swg_index_client.get_entry(data1["baseid"])
-    assert r.did == r2.did
+    res_2 = client.get("/index/" + data1["baseid"])
+    rec_2 = res_2.json
+    assert rec_2["did"] == rec_1["did"]
 
 
-def test_delete_and_recreate(swg_index_client):
+def test_delete_and_recreate(client, user):
     """
     Test that you can delete an IndexDocument and be able to
     recreate it with the same fields.
@@ -462,55 +559,67 @@ def test_delete_and_recreate(swg_index_client):
     new_data = get_doc(has_baseid=True)
     new_data["hashes"] = {"md5": "11111111111111111111111111111111"}
 
-    old_result = swg_index_client.add_entry(old_data)
-    assert old_result.did
-    assert old_result.baseid == old_data["baseid"]
+    old_result = client.post("/index/", json=old_data, headers=user)
+    old_record = old_result.json
+    assert old_record["did"]
+    assert old_record["baseid"] == old_data["baseid"]
 
     # create a new doc with the same did
-    new_data["did"] = old_result.did
+    new_data["did"] = old_record["did"]
 
     # delete the old doc
-    swg_index_client.delete_entry(old_result.did, old_result.rev)
-    with pytest.raises(ApiException):
-        # make sure it's deleted
-        swg_index_client.get_entry(old_result.did)
+    client.delete(
+        "/index/" + old_record["did"] + "?rev=" + old_record["rev"],
+        json=old_data,
+        headers=user,
+    )
+
+    # make sure it's deleted
+    res = client.get("/index/" + old_record["did"])
+    assert res.status_code == 404
 
     # create new doc with the same baseid and did
-    new_result = swg_index_client.add_entry(new_data)
+    new_result = client.post("/index/", json=new_data, headers=user)
+    print(new_result.status_code)
+    new_record = new_result.json
+    print(new_record)
 
-    assert new_result.did
+    assert new_record["did"]
     # verify that they are the same
-    assert new_result.baseid == new_data["baseid"]
-    assert new_result.did == old_result.did
-    assert new_result.baseid == old_result.baseid
+    assert new_record["baseid"] == new_data["baseid"]
+    assert new_record["did"] == old_record["did"]
+    assert new_record["baseid"] == old_record["baseid"]
 
     # verify that new data is in the new node
-    new_doc = swg_index_client.get_entry(new_result.did)
-    assert new_data["baseid"] == new_doc.baseid
-    assert new_data["urls"] == new_doc.urls
-    assert new_data["hashes"]["md5"] == new_doc.hashes.md5
+    new_result = client.get("/index/" + new_record["did"])
+    new_record = new_result.json
+    assert new_data["baseid"] == new_record["baseid"]
+    assert new_data["urls"] == new_record["urls"]
+    assert new_data["hashes"]["md5"] == new_record["hashes"]["md5"]
 
 
-def test_index_create_with_multiple_hashes(swg_index_client):
+def test_index_create_with_multiple_hashes(client, user):
     data = get_doc()
     data["hashes"] = {
         "md5": "8b9942cf415384b27cadf1f4d2d682e5",
         "sha1": "fdbbca63fbec1c2b0d4eb2494ce91520ec9f55f5",
     }
 
-    result = swg_index_client.add_entry(data)
-    assert result.did
+    result = client.post("/index/", json=data, headers=user)
+    record = result.json
+    assert record["did"]
 
 
-def test_index_create_with_valid_did(swg_index_client):
+def test_index_create_with_valid_did(client, user):
     data = get_doc()
     data["did"] = "3d313755-cbb4-4b08-899d-7bbac1f6e67d"
 
-    result = swg_index_client.add_entry(data)
-    assert result.did == "3d313755-cbb4-4b08-899d-7bbac1f6e67d"
+    result = client.post("/index/", json=data, headers=user)
+    record = result.json
+    assert record["did"] == "3d313755-cbb4-4b08-899d-7bbac1f6e67d"
 
 
-def test_index_create_with_acl_authz(swg_index_client):
+def test_index_create_with_acl_authz(client, user):
     data = {
         "acl": ["a", "b"],
         "authz": ["x", "y"],
@@ -520,13 +629,15 @@ def test_index_create_with_acl_authz(swg_index_client):
         "hashes": {"md5": "8b9942cf415384b27cadf1f4d2d682e5"},
     }
 
-    r = swg_index_client.add_entry(data)
-    result = swg_index_client.get_entry(r.did)
-    assert result.acl == ["a", "b"]
-    assert result.authz == ["x", "y"]
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    result = client.get("/index/" + rec["did"])
+    record = result.json
+    assert record["acl"] == ["a", "b"]
+    assert record["authz"] == ["x", "y"]
 
 
-def test_index_create_with_duplicate_acl_authz(swg_index_client):
+def test_index_create_with_duplicate_acl_authz(client, user):
     data = {
         "acl": ["a", "b", "a"],
         "authz": ["x", "y", "x"],
@@ -536,104 +647,115 @@ def test_index_create_with_duplicate_acl_authz(swg_index_client):
         "hashes": {"md5": "8b9942cf415384b27cadf1f4d2d682e5"},
     }
 
-    r = swg_index_client.add_entry(data)
-    result = swg_index_client.get_entry(r.did)
-    assert result.acl == ["a", "b"]
-    assert result.authz == ["x", "y"]
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    result = client.get("/index/" + rec["did"])
+    record = result.json
+    assert record["acl"] == ["a", "b"]
+    assert record["authz"] == ["x", "y"]
 
 
-def test_index_create_with_invalid_did(swg_index_client):
+def test_index_create_with_invalid_did(client, user):
     data = get_doc()
 
     data["did"] = "3d313755-cbb4-4b0fdfdfd8-899d-7bbac1f6e67dfdd"
-
-    with pytest.raises(ApiException) as e:
-        swg_index_client.add_entry(data)
-        assert e.status == 400
+    response = client.post("/index/", json=data, headers=user)
+    assert response.status_code == 400
 
 
-def test_index_create_with_prefix(swg_index_client):
+def test_index_create_with_prefix(client, user):
     data = get_doc()
     data["did"] = "cdis:3d313755-cbb4-4b08-899d-7bbac1f6e67d"
 
-    r = swg_index_client.add_entry(data)
-    assert r.did == "cdis:3d313755-cbb4-4b08-899d-7bbac1f6e67d"
+    response = client.post("/index/", json=data, headers=user)
+    record = response.json
+    assert record["did"] == "cdis:3d313755-cbb4-4b08-899d-7bbac1f6e67d"
 
 
-def test_index_create_with_duplicate_did(swg_index_client):
+def test_index_create_with_duplicate_did(client, user):
     data = get_doc()
     data["did"] = "3d313755-cbb4-4b08-899d-7bbac1f6e67d"
 
-    swg_index_client.add_entry(data)
-
-    with pytest.raises(ApiException) as e:
-        swg_index_client.add_entry(data)
-        assert e.status == 400
+    response = client.post("/index/", json=data, headers=user)
+    response = client.post("/index/", json=data, headers=user)
+    assert response.status_code == 400
 
 
-def test_index_create_with_file_name(swg_index_client):
+def test_index_create_with_file_name(client, user):
     data = get_doc()
     data["file_name"] = "abc"
 
-    r = swg_index_client.add_entry(data)
-    r = swg_index_client.get_entry(r.did)
-    assert r.file_name == "abc"
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    res = client.get("/index/" + rec["did"])
+    rec = res.json
+    assert rec["file_name"] == "abc"
 
 
-def test_index_create_with_version(swg_index_client):
+def test_index_create_with_version(client, user):
     data = get_doc()
     data["version"] = "ver_123"
 
-    r = swg_index_client.add_entry(data)
-    r = swg_index_client.get_entry(r.did)
-    assert r.version == data["version"]
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    res = client.get("/index/" + rec["did"])
+    rec = res.json
+    assert rec["version"] == data["version"]
 
 
-def test_index_create_blank_record(swg_index_client):
+def test_index_create_blank_record(client, user):
     doc = {"uploader": "uploader_123", "baseid": "baseid_123"}
 
-    r = swg_index_client.create_blank_entry(doc)
-    assert r.did
-    res = swg_index_client.get_entry(r.did)
-    assert res.acl == []
-    assert res.authz == []
-    assert res.urls_metadata == {}
-    assert res.size is None
-    assert res.version is None
-    assert res.urls_metadata == {}
+    res = client.post("/index/blank/", json=doc, headers=user)
+    rec = res.json
+    assert rec["did"]
+    res = client.get("/index/" + rec["did"])
+    rec = res.json
+    assert rec["acl"] == []
+    assert rec["authz"] == []
+    assert rec["urls_metadata"] == {}
+    assert rec["size"] is None
+    assert rec["version"] is None
+    assert rec["urls_metadata"] == {}
 
 
-def test_index_create_with_uploader(swg_index_client):
+def test_index_create_with_uploader(client, user):
     data = get_doc()
     data["uploader"] = "uploader_123"
-    r = swg_index_client.add_entry(data)
-    r = swg_index_client.get_entry(r.did)
-    assert r.uploader == data["uploader"]
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    res = client.get("/index/" + rec["did"])
+    rec = res.json
+    assert rec["uploader"] == data["uploader"]
 
 
-def test_index_get_global_endpoint(swg_global_client, swg_index_client):
+def test_index_get_global_endpoint(client, user):
     data = get_doc()
 
-    r = swg_index_client.add_entry(data)
-    r = swg_global_client.get_entry(r.did)
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    res = client.get("/" + rec["did"])
+    rec = res.json
 
-    assert r.metadata == data["metadata"]
-    assert r.form == "object"
-    assert r.size == data["size"]
-    assert r.urls == data["urls"]
-    assert r.hashes.md5 == data["hashes"]["md5"]
+    assert rec["metadata"] == data["metadata"]
+    assert rec["form"] == "object"
+    assert rec["size"] == data["size"]
+    assert rec["urls"] == data["urls"]
+    assert rec["hashes"]["md5"] == data["hashes"]["md5"]
 
-    r2 = swg_global_client.get_entry("testprefix:" + r.did)
-    assert r2.did == r.did
+    res_2 = client.get("/testprefix:" + rec["did"])
+    rec_2 = res_2.json
+    assert rec_2["did"] == rec["did"]
 
 
-def test_index_update(swg_index_client):
+def test_index_update(client, user):
     data = get_doc()
 
-    r = swg_index_client.add_entry(data)
-    assert r.did
-    assert r.rev
-    assert swg_index_client.get_entry(r.did).metadata == data["metadata"]
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    assert rec["did"]
+    assert rec["rev"]
+    assert client.get("/index/" + rec["did"]).json["metadata"] == data["metadata"]
     dataNew = get_doc()
     del dataNew["hashes"]
     del dataNew["size"]
@@ -642,34 +764,43 @@ def test_index_update(swg_index_client):
     dataNew["version"] = "ver123"
     dataNew["acl"] = ["a", "b"]
     dataNew["authz"] = ["x", "y"]
-    r2 = swg_index_client.update_entry(r.did, rev=r.rev, body=dataNew)
-    assert r2.rev != r.rev
-    result = swg_index_client.get_entry(r.did)
-    assert result.metadata == dataNew["metadata"]
-    assert result.acl == dataNew["acl"]
-    assert result.authz == dataNew["authz"]
+    res_2 = client.put(
+        "/index/" + rec["did"] + "?rev=" + rec["rev"], json=dataNew, headers=user
+    )
+    rec_2 = res_2.json
+    assert rec_2["rev"] != rec["rev"]
+    response = client.get("/index/" + rec_2["did"])
+    record = response.json
+    assert record["metadata"] == dataNew["metadata"]
+    assert record["acl"] == dataNew["acl"]
+    assert record["authz"] == dataNew["authz"]
 
     data = get_doc()
     data["did"] = "cdis:3d313755-cbb4-4b08-899d-7bbac1f6e67d"
-    r = swg_index_client.add_entry(data)
-    assert r.did
-    assert r.rev
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    assert rec["did"]
+    assert rec["rev"]
     dataNew = {
         "urls": ["s3://endpointurl/bucket/key"],
         "file_name": "test",
         "version": "ver123",
     }
-    r2 = swg_index_client.update_entry(r.did, rev=r.rev, body=dataNew)
-    assert r2.rev != r.rev
+    res_2 = client.put(
+        "/index/" + rec["did"] + "?rev=" + rec["rev"], json=dataNew, headers=user
+    )
+    rec_2 = res_2.json
+    assert rec_2["rev"] != rec["rev"]
 
 
-def test_index_update_duplicate_acl_authz(swg_index_client):
+def test_index_update_duplicate_acl_authz(client, user):
     data = get_doc()
 
-    r = swg_index_client.add_entry(data)
-    assert r.did
-    assert r.rev
-    assert swg_index_client.get_entry(r.did).metadata == data["metadata"]
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    assert rec["did"]
+    assert rec["rev"]
+    assert client.get("/index/" + rec["did"]).json["metadata"] == data["metadata"]
     dataNew = get_doc()
     del dataNew["hashes"]
     del dataNew["size"]
@@ -678,61 +809,78 @@ def test_index_update_duplicate_acl_authz(swg_index_client):
     dataNew["version"] = "ver123"
     dataNew["acl"] = ["c", "d", "c"]
     dataNew["authz"] = ["x", "y", "x"]
-    r2 = swg_index_client.update_entry(r.did, rev=r.rev, body=dataNew)
-    assert r2.rev != r.rev
-    result = swg_index_client.get_entry(r.did)
-    assert result.metadata == dataNew["metadata"]
-    assert result.acl == ["c", "d"]
-    assert result.authz == ["x", "y"]
+    res_2 = client.put(
+        "/index/" + rec["did"] + "?rev=" + rec["rev"], json=dataNew, headers=user
+    )
+    rec_2 = res_2.json
+    assert rec_2["rev"] != rec["rev"]
+    response = client.get("/index/" + rec["did"])
+    record = response.json
+    assert record["metadata"] == dataNew["metadata"]
+    assert record["acl"] == ["c", "d"]
+    assert record["authz"] == ["x", "y"]
 
 
-def test_update_uploader_field(swg_index_client):
+def test_update_uploader_field(client, user):
     data = get_doc()
     data["uploader"] = "uploader_123"
-    r = swg_index_client.add_entry(data)
-    assert r.did
-    assert r.rev
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    assert rec["did"]
+    assert rec["rev"]
 
-    r = swg_index_client.get_entry(r.did)
-    assert r.uploader == "uploader_123"
+    res = client.get("/index/" + rec["did"])
+    rec = res.json
+    assert rec["uploader"] == "uploader_123"
 
     updated = {"uploader": "new_uploader"}
-    swg_index_client.update_entry(r.did, rev=r.rev, body=updated)
+    client.put(
+        "/index/" + rec["did"] + "?rev=" + rec["rev"], json=updated, headers=user
+    )
 
-    r = swg_index_client.get_entry(r.did)
-    assert r.uploader == "new_uploader"
+    res = client.get("/index/" + rec["did"])
+    rec = res.json
+    assert rec["uploader"] == "new_uploader"
 
     updated = {"uploader": None}
-    swg_index_client.update_entry(r.did, rev=r.rev, body=updated)
+    client.put(
+        "/index/" + rec["did"] + "?rev=" + rec["rev"], json=updated, headers=user
+    )
 
-    r = swg_index_client.get_entry(r.did)
-    assert r.uploader is None
+    res = client.get("/index/" + rec["did"])
+    rec = res.json
+    assert rec["uploader"] is None
 
 
-def test_index_delete(swg_index_client):
+def test_index_delete(client, user):
     data = get_doc(has_metadata=False, has_baseid=False)
 
-    r = swg_index_client.add_entry(data)
-    assert r.did
-    assert r.rev
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    assert rec["did"]
+    assert rec["rev"]
 
-    r = swg_index_client.get_entry(r.did)
-    assert r.did
+    res = client.get("/index/" + rec["did"])
+    rec = res.json
+    assert rec["did"]
 
-    swg_index_client.delete_entry(r.did, rev=r.rev)
+    client.delete(
+        "/index/" + rec["did"] + "?rev=" + rec["rev"], json=data, headers=user
+    )
 
-    with pytest.raises(ApiException) as e:
-        r = swg_index_client.get_entry(r.did)
-        assert e.status == 400
+    # make sure its deleted
+    res = client.get("/index/" + rec["did"] + "?rev=" + rec["rev"])
+    assert res.status_code == 404
 
 
-def test_create_index_version(swg_index_client):
+def test_create_index_version(client, user):
     data = get_doc(has_metadata=False, has_baseid=False)
 
-    r = swg_index_client.add_entry(data)
-    assert r.did
-    assert r.rev
-    assert r.baseid
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    assert rec["did"]
+    assert rec["rev"]
+    assert rec["baseid"]
 
     dataNew = {
         "did": "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
@@ -743,44 +891,53 @@ def test_create_index_version(swg_index_client):
         "acl": ["a"],
     }
 
-    r2 = swg_index_client.add_new_version(r.did, body=dataNew)
-    assert r2.baseid == r.baseid
-    assert r2.did == dataNew["did"]
+    res_2 = client.post("/index/" + rec["did"], json=dataNew, headers=user)
+    rec_2 = res_2.json
+    assert rec_2["baseid"] == rec["baseid"]
+    assert rec_2["did"] == dataNew["did"]
 
 
-def test_get_latest_version(swg_index_client):
+def test_get_latest_version(client, user):
     data = get_doc(has_metadata=False, has_baseid=False, has_version=True)
-    r = swg_index_client.add_entry(data)
-    assert r.did
+    res_1 = client.post("/index/", json=data, headers=user)
+    rec_1 = res_1.json
+    assert rec_1["did"]
 
     data = get_doc(has_metadata=False, has_baseid=False, has_version=False)
-    r2 = swg_index_client.add_new_version(r.did, body=data)
-    r3 = swg_index_client.get_latest_version(r.did)
-    assert r3.did == r2.did
+    res_2 = client.post("/index/" + rec_1["did"], json=data, headers=user)
+    rec_2 = res_2.json
+    res_3 = client.get("/index/" + rec_2["did"] + "/latest")
+    rec_3 = res_3.json
+    assert rec_3["did"] == rec_2["did"]
 
-    r4 = swg_index_client.get_latest_version(r.baseid)
-    assert r4.did == r2.did
+    res_4 = client.get("/index/" + rec_1["baseid"] + "/latest")
+    rec_4 = res_4.json
+    assert rec_4["did"] == rec_2["did"]
 
-    r5 = swg_index_client.get_latest_version(r.baseid, has_version=True)
-    assert r5.did == r.did
+    res_5 = client.get("/index/" + rec_1["baseid"] + "/latest?has_version=True")
+    rec_5 = res_5.json
+    assert rec_5["did"] == rec_1["did"]
 
 
-def test_get_all_versions(swg_index_client):
+def test_get_all_versions(client, user):
     data = get_doc(has_metadata=False, has_baseid=False)
-    r = swg_index_client.add_entry(data)
-    assert r.did
-    swg_index_client.add_new_version(r.did, body=data)
-    r3 = swg_index_client.get_all_versions(r.did)
-    assert len(r3) == 2
-    r4 = swg_index_client.get_all_versions(r.baseid)
-    assert len(r4) == 2
+    res_1 = client.post("/index/", json=data, headers=user)
+    rec_1 = res_1.json
+    assert rec_1["did"]
+    client.post("/index/" + rec_1["did"], json=data, headers=user)
+    res_2 = client.get("/index/" + rec_1["did"] + "/versions")
+    rec_2 = res_2.json
+    assert len(rec_2) == 2
+    res_3 = client.get("/index/" + rec_1["baseid"] + "/versions")
+    rec_3 = res_3.json
+    assert len(rec_3) == 2
 
 
-def test_alias_list(swg_alias_client):
-    assert swg_alias_client.list_entries().aliases == []
+def test_alias_list(client, user):
+    assert client.get("/alias/").json["aliases"] == []
 
 
-def test_alias_create(swg_alias_client):
+def test_alias_create(client, user):
     data = {
         "size": 123,
         "hashes": {"md5": "8b9942cf415384b27cadf1f4d2d682e5"},
@@ -789,29 +946,15 @@ def test_alias_create(swg_alias_client):
         "host_authorities": ["PDC"],
     }
     ark = "ark:/31807/TEST-abc"
-    r = swg_alias_client.upsert_entry(ark, body=data)
-    assert r.name == ark
+    res = client.put("/alias/" + ark, json=data, headers=user)
+    rec = res.json
+    assert rec["name"] == ark
 
-    assert len(swg_alias_client.list_entries().aliases) == 1
-    assert swg_alias_client.get_entry(r.name).name
-
-
-def test_alias_get_global_endpoint(swg_alias_client, swg_global_client):
-    data = {
-        "size": 123,
-        "hashes": {"md5": "8b9942cf415384b27cadf1f4d2d682e5"},
-        "release": "private",
-        "keeper_authority": "CRI",
-        "host_authorities": ["PDC"],
-    }
-    ark = "ark:/31807/TEST-abc"
-
-    swg_alias_client.upsert_entry(ark, body=data)
-
-    assert swg_global_client.get_entry(ark).size == 123
+    assert len(client.get("/alias/").json["aliases"]) == 1
+    assert client.get("/alias/" + rec["name"]).json["name"]
 
 
-def test_alias_update(swg_alias_client):
+def test_alias_get_global_endpoint(client, user):
     data = {
         "size": 123,
         "hashes": {"md5": "8b9942cf415384b27cadf1f4d2d682e5"},
@@ -821,8 +964,24 @@ def test_alias_update(swg_alias_client):
     }
     ark = "ark:/31807/TEST-abc"
 
-    r = swg_alias_client.upsert_entry(ark, body=data)
-    assert r.rev
+    res = client.put("/alias/" + ark, json=data, headers=user)
+
+    assert client.get("/" + ark).json["size"] == 123
+
+
+def test_alias_update(client, user):
+    data = {
+        "size": 123,
+        "hashes": {"md5": "8b9942cf415384b27cadf1f4d2d682e5"},
+        "release": "private",
+        "keeper_authority": "CRI",
+        "host_authorities": ["PDC"],
+    }
+    ark = "ark:/31807/TEST-abc"
+
+    res_1 = client.put("/alias/" + ark, json=data, headers=user)
+    rec_1 = res_1.json
+    assert rec_1["rev"]
 
     dataNew = {
         "size": 456,
@@ -831,11 +990,14 @@ def test_alias_update(swg_alias_client):
         "keeper_authority": "CRI",
         "host_authorities": ["PDC"],
     }
-    r2 = swg_alias_client.upsert_entry(ark, rev=r.rev, body=dataNew)
-    assert r2.rev != r.rev
+    res_2 = client.put(
+        "/alias/" + ark + "?rev=" + rec_1["rev"], json=dataNew, headers=user
+    )
+    rec_2 = res_2.json
+    assert rec_2["rev"] != rec_1["rev"]
 
 
-def test_alias_delete(swg_alias_client):
+def test_alias_delete(client, user):
     data = {
         "size": 123,
         "hashes": {"md5": "8b9942cf415384b27cadf1f4d2d682e5"},
@@ -845,12 +1007,13 @@ def test_alias_delete(swg_alias_client):
     }
     ark = "ark:/31807/TEST-abc"
 
-    r = swg_alias_client.upsert_entry(ark, body=data)
-    assert r.rev
+    res_1 = client.put("/alias/" + ark, json=data, headers=user)
+    rec_1 = res_1.json
+    assert rec_1["rev"]
 
-    swg_alias_client.delete_entry(ark, rev=r.rev)
+    client.delete("/alias/" + ark + "?rev=" + rec_1["rev"], json=data, headers=user)
 
-    assert len(swg_alias_client.list_entries().aliases) == 0
+    assert len(client.get("/alias/").json["aliases"]) == 0
 
 
 @pytest.mark.parametrize(
@@ -928,77 +1091,106 @@ def test_bad_hashes(client, user, typ, h):
         assert "does not match" in json_resp["error"]
 
 
-def test_dos_get(swg_index_client, swg_dos_client):
+def test_dos_get(client, user):
     data = get_doc(has_urls_metadata=True, has_metadata=True, has_baseid=True)
 
-    result = swg_index_client.add_entry(data)
-    r = swg_dos_client.get_data_object(result.did)
-    assert r.data_object.id == result.did
-    assert r.data_object.size == "123"
-    assert r.data_object.checksums[0].checksum == "8b9942cf415384b27cadf1f4d2d682e5"
-    assert r.data_object.checksums[0].type == "md5"
-    assert r.data_object.urls[0].url == "s3://endpointurl/bucket/key"
-    assert r.data_object.urls[0].user_metadata["state"] == "uploaded"
-    assert r.data_object.urls[0].system_metadata["project_id"] == "bpa-UChicago"
-    r2 = swg_dos_client.get_data_object(result.baseid)
-    assert r2.data_object.id == result.did
+    res_1 = client.post("/index/", json=data, headers=user)
+    rec_1 = res_1.json
+    res_2 = client.get("/ga4gh/dos/v1/dataobjects/" + rec_1["did"])
+    rec_2 = res_2.json
+    assert rec_2["data_object"]["id"] == rec_1["did"]
+    assert rec_2["data_object"]["size"] == 123
+    assert (
+        rec_2["data_object"]["checksums"][0]["checksum"]
+        == "8b9942cf415384b27cadf1f4d2d682e5"
+    )
+    assert rec_2["data_object"]["checksums"][0]["type"] == "md5"
+    assert rec_2["data_object"]["urls"][0]["url"] == "s3://endpointurl/bucket/key"
+    assert rec_2["data_object"]["urls"][0]["user_metadata"]["state"] == "uploaded"
+    assert (
+        rec_2["data_object"]["urls"][0]["system_metadata"]["project_id"]
+        == "bpa-UChicago"
+    )
+    res_3 = client.get("/ga4gh/dos/v1/dataobjects/" + rec_1["baseid"])
+    rec_3 = res_3.json
+    assert rec_3["data_object"]["id"] == rec_1["did"]
 
 
-def test_dos_list(swg_index_client, swg_dos_client):
+def test_dos_list(client, user):
     data = get_doc(has_urls_metadata=True, has_metadata=True, has_baseid=True)
 
-    result = swg_index_client.add_entry(data)
-    r = swg_dos_client.list_data_objects(page_size=100)
-    assert len(r.data_objects) == 1
-    assert r.data_objects[0].id == result.did
-    assert r.data_objects[0].size == "123"
-    assert r.data_objects[0].checksums[0].checksum == "8b9942cf415384b27cadf1f4d2d682e5"
-    assert r.data_objects[0].checksums[0].type == "md5"
-    assert r.data_objects[0].urls[0].url == "s3://endpointurl/bucket/key"
-    assert r.data_objects[0].urls[0].user_metadata["state"] == "uploaded"
-    assert r.data_objects[0].urls[0].system_metadata["project_id"] == "bpa-UChicago"
+    res_1 = client.post("/index/", json=data, headers=user)
+    rec_1 = res_1.json
+
+    res_2 = client.get("/ga4gh/dos/v1/dataobjects?page_size=100")
+    rec_2 = res_2.json
+    assert len(rec_2["data_objects"]) == 1
+    assert rec_2["data_objects"][0]["id"] == rec_1["did"]
+    assert rec_2["data_objects"][0]["size"] == 123
+    assert (
+        rec_2["data_objects"][0]["checksums"][0]["checksum"]
+        == "8b9942cf415384b27cadf1f4d2d682e5"
+    )
+    assert rec_2["data_objects"][0]["checksums"][0]["type"] == "md5"
+    assert rec_2["data_objects"][0]["urls"][0]["url"] == "s3://endpointurl/bucket/key"
+    assert rec_2["data_objects"][0]["urls"][0]["user_metadata"]["state"] == "uploaded"
+    assert (
+        rec_2["data_objects"][0]["urls"][0]["system_metadata"]["project_id"]
+        == "bpa-UChicago"
+    )
 
 
-def test_update_without_changing_fields(swg_index_client):
+def test_update_without_changing_fields(client, user):
     # setup test
     data = get_doc(has_urls_metadata=True, has_metadata=True, has_baseid=True)
 
-    result = swg_index_client.add_entry(data)
-    first_doc = swg_index_client.get_entry(result.did)
+    res = client.post("/index/", json=data, headers=user)
+    rec = res.json
+    first_doc = client.get("/index/" + rec["did"]).json
 
     # update
     updated = {"version": "at least 2"}
-    swg_index_client.update_entry(first_doc.did, rev=first_doc.rev, body=updated)
+    client.put(
+        "/index/" + first_doc["did"] + "?rev=" + first_doc["rev"],
+        json=updated,
+        headers=user,
+    )
 
     # Check if update successful.
-    second_doc = swg_index_client.get_entry(first_doc.did)
+    second_doc = client.get("/index/" + first_doc["did"]).json
     # Only `version` changed.
-    assert first_doc.version != second_doc.version
+    assert first_doc["version"] != second_doc["version"]
 
     # The rest is the same.
-    assert first_doc.urls == second_doc.urls
-    assert first_doc.size == second_doc.size
-    assert first_doc.file_name == second_doc.file_name
-    assert first_doc.metadata == second_doc.metadata
+    assert first_doc["urls"] == second_doc["urls"]
+    assert first_doc["size"] == second_doc["size"]
+    assert first_doc["file_name"] == second_doc["file_name"]
+    assert first_doc["metadata"] == second_doc["metadata"]
 
     # Change `version` to null.
     # update
     updated = {"version": None}
-    swg_index_client.update_entry(second_doc.did, rev=second_doc.rev, body=updated)
+    client.put(
+        "/index/" + second_doc["did"] + "?rev=" + second_doc["rev"],
+        json=updated,
+        headers=user,
+    )
 
     # check if update successful
-    third_doc = swg_index_client.get_entry(result.did)
+    third_doc = client.get("/index/" + rec["did"]).json
     # Only `version` changed.
-    assert second_doc.version != third_doc.version
+    assert second_doc["version"] != third_doc["version"]
 
 
-def test_bulk_get_documents(swg_index_client, swg_bulk_client):
+def test_bulk_get_documents(client, user):
 
     # just make a bunch of entries in indexd
-    dids = [swg_index_client.add_entry(get_doc(has_baseid=True)).did for _ in range(20)]
-
+    dids = [
+        client.post("/index/", json=get_doc(has_baseid=True), headers=user).json["did"]
+        for _ in range(20)
+    ]
     # do a bulk query for them all
-    docs = swg_bulk_client.get_bulk_ids(dids)
+    docs = client.post("/bulk/documents", json=dids, headers=user).json
 
     # compare that they are the same by did
     for doc in docs:
